feat(interfaces): Export IFulltextSearchParams and ILabelSearchParams
