+++ src/interfaces/models/v2/i-fulltext-search-params.ts
@@ -1,0 +1,21 @@
+/**
+ * Params for fulltext search.
+ */
+export interface IFulltextSearchParams {
+
+    /**
+     * Iri of resource class the fulltext search is restricted to, if any.
+     */
+    limitToResourceClass?: string;
+
+    /**
+     * Iri of the project the fulltext search is restricted to, if any.
+     */
+    limitToProject?: string;
+
+    /**
+     * Iri of standoff class the fulltext search is restricted to, if any.
+     */
+    limitToStandoffClass?: string;
+
+}
