+++ src/interfaces/models/v2/i-label-search-params.ts
@@ -1,0 +1,16 @@
+/**
+ * Params for search by label
+ */
+export interface ILabelSearchParams {
+
+    /**
+     * Iri of the project the search by label is restricted to, if any.
+     */
+    limitToResourceClass?: string;
+
+    /**
+     * Iri of the project the search by label is restricted to, if any.
+     */
+    limitToProject?: string;
+
+}
